refactor: BaseFragment 패키지 조정
- 공통으로 사용될 것 같으므로 이를 조정함
