+++ app/src/main/java/com/capstone/artwhale/presentation/home/profile/my/MyArtFragment.kt
@@ -4,7 +4,7 @@ import android.content.Context
 import androidx.activity.OnBackPressedCallback
 import com.capstone.artwhale.R
 import com.capstone.artwhale.databinding.FragmentMyArtBinding
-import com.capstone.artwhale.presentation.home.BaseFragment
+import com.capstone.artwhale.presentation.common.BaseFragment
 import com.capstone.artwhale.presentation.home.profile.common.adapter.ListVPAdapter
 import com.google.android.material.tabs.TabLayoutMediator
