+++ app/src/main/java/com/capstone/artwhale/presentation/register/music/ai/AIAlbumSelectFragment.kt
@@ -1,7 +1,7 @@
 package com.capstone.artwhale.presentation.register.music.ai
 
 import com.capstone.artwhale.databinding.FragmentAiAlbumSelectBinding
-import com.capstone.artwhale.presentation.home.BaseFragment
+import com.capstone.artwhale.presentation.common.BaseFragment
 
 class AIAlbumSelectFragment :
     BaseFragment<FragmentAiAlbumSelectBinding>(FragmentAiAlbumSelectBinding::inflate) {
