+++ app/src/main/java/com/capstone/artwhale/presentation/register/music/upload/MusicUploadFragment.kt
@@ -1,7 +1,7 @@
 package com.capstone.artwhale.presentation.register.music.upload
 
 import com.capstone.artwhale.databinding.FragmentMusicUploadBinding
-import com.capstone.artwhale.presentation.home.BaseFragment
+import com.capstone.artwhale.presentation.common.BaseFragment
 
 class MusicUploadFragment :
     BaseFragment<FragmentMusicUploadBinding>(FragmentMusicUploadBinding::inflate) {
